Moved ComplicationArray classes to separate file
